backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langchain-community>=0.0.1
langchain-ollama>=0.1.0
chromadb>=0.4.0
ollama>=0.1.0
python-dotenv>=0.19.0
requests>=2.26.0
pydantic-settings>=2.0.0
